+++ CHANGELOG.md
@@ -21,13 +21,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [5.1.5]
 
 ### Added
+* tests for LinkFactory
 
 ### Changed
 
 ### Deprecated
 
 ### Fixed
-* fix pad url parsing when server is using proxy
+* fix bad url parsing when server is using proxy
 
 ### Removed